23041 refactor(RichText): adopt BEM classes and improve truncation UX
Restructure checklist, progressbar, and collapse styles with BEM naming.
Move RichTextMoreLink to components/, replace anchor with button, and fix
more-link overlay positioning for line-clamped content.
