+++ frontend/src/public/components/TaskCard/checklist/TaskCheckableItem/TaskCheckableItem.tsx
@@ -49,7 +49,6 @@ export function TaskCheckableItem({
     <Checkbox
       checked={isChecked}
       containerClassName={styles['checkbox']}
-      labelClassName={styles['checkbox-label']}
       onChange={handleChange}
       disabled={disabled}
     />,
